@@ -13,7 +13,7 @@ object Core {
         }
     }
 
-    val ARG_REPORT_ID = "id"
+    const val ARG_REPORT_ID = "id"
     val mi: MediaInfo = MediaInfo()
     val views: MutableList<ReportView> = mutableListOf()
     val version: String = mi.Option("Info_Version").replace("MediaInfoLib - v", "")
